T44 Merge in `fastapi-demo` and `flask-demo`
Merge in the FastAPI and Flask application blueprints from the
`fastapi-demo` and `flask-demo` repositories, which can now be
archived.

Also rename this project from `lambda-boilerplate` to the more fitting
`serverless-blueprints`, now that it contains several different
Serverless blueprints.

Bump dependency versions and align the code between the different
blueprints better while we're at it.